Proof of concept for TUI
This TUI demo builds a somewhat generlised input/output pair with tab
cycling for the input components.

Here we build four 64 bit integer input fields. Writing a value to any
of these fields will reflect the edited values in the output fields.

Eventually we should be able to build a range of these fields which all
represent a different view on a single register for an instruction.

For example a 256 bit AVX2 register can be thought of as four 64 bit
values, 8 32 bit values, 16 16 bit values or 32 8 bit values. We intend
to have this one register able to be edited from any one of those
different _perspectives_. Editing the value in one perspective will
update the input/output values for all other perspectives. All of these
perspectives represent different views on a single common register.

Another perspective which is important for us is to allow the
input/output to use different base systems to display and edit the
register values. Base 10 is supported now, but base 16 is likely more
useful in practice and maybe octal and binary should be valid options
too. The base 10 nature of the input/output groups is hard coded right
now but will be generalised in the future.