Fix infinite update loop
There was a bug where an input is changed triggering the InputRegisters
to re-calculate the register value and broadcast this change to all
receivers. One of the receivers was the InputRegisters which originated
the change. The originating input would be updated with the value it had
just changed to, and this would trigger _another_ update event causing
the input to be updated again. This is an recursive infinite loop that
only ends when the stack overflows its size limit.

There are two fixes here. One is the give each change receiver a UUID,
and add that UUID to the data-changed broadcast. Then when updating
don't update an input if the InputRegister's id matches the data-changed
sender's id. Problem solved?

Unfortunately in the future we will have multiple InputRegisters who
both receive data-changed broadcasts and send data-changed broadcasts on
update. Using only the UUID solution we would have one InputRegister
send a data-changed broadcast which would trigger an update on another
InputRegister who would in turn broadcast their own data-change
broadcast which would update the first InputRegister. So we now have
mutual recursion, more subtle but equally infinite.

The solution then is to _only_ update an input-field if the text of the
input-field is different from the change text.

We have kept the UUIDs although they are not used.