Allow change broadcaster to be de/reactivated
There is a bug caused by having the field updates caused by data-changed
events generating their own data-changed events which are then further
broadcast. So far we have tried several approaches to resolve this
problem.

1: Set an id on broadcast data-changed events. Don't consume your own
   data-changed events. This prevents a data-change infinite loop, but
other problems remain.
2: Only allow an input/output to be updated if the data being changed
   actually changes the value stored in the field.

There was a new bug which caused a kind of unbounded update loop which
would grow the value being set at each iteration. This would eventually
cause a panic when the value being updated was too large for the bitsize
of the field being set.

The actual cause was never fully understood.

However, simply preventing all field updates which happen in response to
a data-changed event broadcast resolves all of these problems. We allow
the RegisterInputs to deactivate the changeBroadcaster while it updates
it's inputs in response to a data-changed event. Now no ping-pong
update broadcasts are possible. This also dramatically simplifies the
logs since events are only broadcast once.

There may be a bug lurking in the code as the true cause of the growing
update value is not known right now. Hopefully we will understand this
in the future.