changeBroadcaster now sets values for ui 'parts'
Previously the changeBroadcaster would provide the full register of
bytes to the ui component to update itself with. This created two
methods on RegisterInputs and RegisterOutputs which were almost
identical and weren not short. I had found myself repeatedly updating
the code in one and then having to go and repeat the update for the
other.

This code change now means that both RegisterInputs and RegisterOutputs
satisfy the following interface

       dataParts() int
       setPart(i int, chunk []byte)
       describe() string

The changeBroadcaster then splits the update bytes evenly across the
number of 'data-parts' and sets each part in a loop.

An alternative approach would be to factor out a common type for
RegisterInputs/RegisterOutputs which performed this processing. This is
doable, but the fact that setting text for a tview.InputField is
different from a tview.TextView would require a set of wrapper types and
would introduce at least three new types plus a (imo) confusing level of
indirection. I would have preferred to unify the RegisterInput/Output
types, but couldn't find a way to do it that didn't feel worse than the
solution chosen in this commit.