Unified the register-input/out types
The unified type is RegisterParts. This introduces a new interface to
abstract over the InputField and the TextViews. This complicating change
feels necessary because the constructor for both RegisterInput and
Output was identical (more or less) and becoming more and more complex.

The change which prompted this was adding borders and titles to the
input/output parts as well as arranging them across multiple lines. It
became clear that a 256 bit register could not be represented as 88 bit
values on a single line in my terminal. So we split it up into multple
lines with at most 8 parts per line. We add a title, plus border) to
each part so we could clearly write down the bit offset ranges for each
part. I do not want to write this kind of code twice and in particular
don't want to have to fiddle the UI by updating two separate but almost
identical constructors.

One or two commits ago I described this approach as a possible, but
rejected, approach for resolving the duplication of the broadcast
updates code. Since more duplication has forced our hand here we can
probably revisit that change.