Move the register ui components into register pkg
We wrap the whole UI component into a new register.UIRegisterSet type.
This contains a set of input/output pairs all linked together by a
change-broadcaster. The register set has three UIRegister fields, one
for each of the base systems 2, 10 and 16.

The intention is that these are all built and linked together and the
user can switch between different base systems, and we will simply
swich out the UIRegister being used to actually render the UI.