Calculate 'parts per line' dynamically
Previously we would try to display 8 register parts per line. But this
approach breaks down spectaculaly when we have base 2 parts or try to
render for a 512 bit register.

Now we calculate the width of the screen, then calculate the width of
each part and work out how many to display per line based on that. Not
thoroughly tested, but seems to work reasonably well.