Add stackapp package
This package wraps a tview.Application and allows us to manage a stack
based navigation scheme for ui screens.

Very first steps here, currently pop wipes out the popped screen, but
probably it will be useful to keep this to provide a 'forward'
navigation step, similar to what we have in a browser. Time will tell.