Use stackapp and instructions pkgs in commands
We replace the previous slice of strings with a slice of
instructions.Instruction. This means that our CommandSearch UI screen
now manages a collection of instructions.

The first thing that jumps out is that our naming is now all over the
place. Instructions vs commands, I bet there are more inconsistencies.
Will be cleaned up.