UIRegister now models input->instruction->output
Previously UIRegister modelled a single SIMD register which could be
modified via inputs and have those changes reflected in the outputs.
Fine.

However, not all SIMD instructions look like modifying a large SIMD
register. For example VPBROADCASTB is best modelled as having the input
of a single byte and the output is a large SIMD register, in our
currently code base we use a 256 bit register.

To model this more effectively, we split the inputs and outputs into
different sizes. The change broadcaster is now in charge of taking new
inputs running the instruction and then broadcasting the updated
inputs/outputs to the registered input/output receivers.

This has caused some more renaming, and (as mentioned in previous
commits) I think many names in the codebase are innapropriate. I've made
a reasonable effort to fix up the ones which were obvious to me while
working. But a dedicated naming review is needed in the near future.
Likely when everything is running and we can demonstrate at least one
functional instruction.