Split assembly functions into separate files
This allows us to embed the generated assembly code (and stubs) directly
into the instruction struct and display it in the UI.

We also add a Supported() bool to the Instruction interface. This allows
us to handle unsupported instructions, for example on the computer I am
working the 512 bit version of VPBROADCASTB is not supported, and so
cannot be demonstrated.