Use regex to extract CPU features
Previously we used strings.Contains(...) to find lines with CPU features
in the assembly. My messy approach produced CPU feature strings with
spaces and trailing commas as well as sometimes just the empty string
"". Now we use a regex to cleanly extract the prefix and more carefully
split the remaining features so they are perfect when hasFeature(...)
receives them.

Possibly not a big deal, but it bothered me so now it's fixed.