Adding number package
This package implements the Converter interface as well as providing a
range of static helper functions converting number types to and from
[]byte.

This replaces the valueConverter struct that previously lived inside the
register package. We needed a Converter interface in order to more
tidily manage converting between int/uint/float representations with
different integer bases (floats are _always_ base 10) and for different
bith widths. With three dimensions I didn't want to just use one struct
which tried to do everything (although that _is_ how this change
started).

This change became rather viral, as once to decide that _maybe_ a field
might be a float or an integer, with different sizes and base systems,
you need a value carrying all this information to be introduced near the
beginning as we try to assemble the UI components.

In order to do this I tried switching the assembly.Instruction
interface's InputSizes method to hold a slice of converters, instead of
just a slice of input sizes. This sort-of works, and can be improved but
we immediately hit a fundamental flaw which is that the converters are
_really_ targeting a single input field, i.e. if you have 256 bit
register input and the UI models it as four 64 bit input integers, the
Converter lives at the 64 bit input field level - but the instruction's
InputSizes field is describing the whole input size (in this example 256
bit). So we need some way to manage this, maybe by allowing the
converters to split themselves into smaller parts?

This commit compiles and runs, right up when you choose an instruction
to display. Then it crashes because we ask it to build an input field
for a 256 bit register (if you chose the VPBROADCAST256), which is
considered an erroneous part size. The next commit will try to resolve
this new problem?

Should we be trying to push this complexity into the Instruction
interface (that is where some of this complexity has been pushed in this
commit)? This choice feels natural in that each instruction can be
expected to have an opinion on whether it's arguments are integers or
floats. But the splitting args down doesn't seem to sit naturally with
the Instruction interface. Maybe it does, if you have a pair of 256 bit
registers but the instruction is summing 4 64 bit float pairs, then
splitting to 64 bit pieces makes sense and is aligned with the
instruction itself.