Introduce the number.Parameter type
This type describes a parameter, input or output, for an instruction.
This includes the bit-width the text-width the base system and a
converter for doing []byte<->string conversions.

We fixed the base of all inputs/outputs, where previously we had thought
that users would want the freedom to switch between different base
representations. Looking at a handful of instructions this doesn't
actually seem that useful, for example if you are adding two sets of
32 bit floats, then you just want everything in 32 bit float. If you are
using a write mask, you really only want that write mask in base16. So
we've pushed this information into the assembly.Instruction itself and
made it fixed. Maybe this will be changed in the future, but right now I
don't expect that to happen.