register package now uiio
This stands for user-interface input-output.

Originally there was a concept where the UI was imagined to be a way to
'set a register' for the instruction. This makes sense, but really the
instructions being exercised are contained in assembly functions which
take arguments, named parameters in our code base. These ui components
are really modifying the value of the parameters passed into the
assembly functions, so it feels more natural to name them along these
lines.

So the register package now has a snappy new name and all the types are
renamed.

UIRegisterSet -> UIParametersSet
UIRegister    -> UIParameters
RegisterParts -> UIParameterParts
uiPart        -> uiParameterPart

One thing that I noticed while reviewing this commit was that if we
really do remove the UIParametersSet, which _was_ a container for
different base system representations of all parameters, the
changeBroadcaster doesn't need to update the input parameter fields
anymore - because there's only one set of input parameter fields. Only
the output needs to be updated via the broadcast. This removes the
tricky recursive update problem that we had previously dealt with. Could
be nice.