Adding (v)addps instructions
This is an important step because this instruction requires two
arguments (size 128 or 256 bits).

We _do_ generate the two parameters of the correct size, but it's not
possible to actually select the second parameter. We can't tab to it,
the first parameter is selected by default and tabbing only cycles
through fields for the first parameter.

Mouse selection doesn't seem to work at all. It's not clear to me why
this doesn't work.