Eliminate changeBroacaster
Previously we used this to synchronise all the different input/output
sets with different base systems. Now we fix the base system for a given
instruction and don't allow arbitrary changes to it. So we don't need
this somewhat complex synchronisation system.

Also renamed UIParameters to UIInstruction, a much clearer name.