Fix vmovdqa package name was erroneously vmovdqu
Fixing this allows us to access these instructions in all.Instructions()

There's a subtle lesson in this commit. I hadn't noticed that these
instructions had an erroneous package name, nor had I noticed that they
were missing from all.Instructions(). Right now we have only a
relatively small set of instructions, and already it is getting hard to
manage by hand. The solution to this will likely lie in a number of
separate small fixes which simply reduce the manual management required
for the set of instructions. No immediate action is required, but we
should be moving constantly toward reducing the burden of this
management.