Detect unstable float inputs and set input red
When the user enters a floating point string value which parses to a different
value when converting to a float32/64 we set the input to red. This is
intended to warn the user to expect a slightly different output from
what they might expect if a perfect float translation was possible.

In the future we will likely need to include some help text or similar
to explain this behaviour. But not now, that is very much a refinement.