Adding movhps instructions
I realised looking at these instructions, that the movhps isntruction
only really makes sense in a merge context. i.e. that we are taking a
register with values populated in the lower half and we are now loading
the upper half. The vmovhps version just allows us to choose a different
register for the output of this operation.

We need to go back and make the movlps instructions demonstrate this
merge behaviour.