movhps_64_loadmergestore.go template generated
The avo assembly generator, stubs assembly and the type itself were
generated from a template. To be clear the template generates a dummy
implementation stub for these files, so the implementation details had
to be manually preserved through the re-generation process. However,
this does rename the assembly function and the go type. We will now go
through all of the assembly instructions and regenerate them in the same
way.

After that I would actually like to rebuild the file naming convention
too. Should we do this now, or in two steps. It does feel like double
passing to change the file name convention _after_ regenerating the
internal names. But I think trying to do both at once will dramatically
increase the rate of errors. Doing it in two passes.