Adding register useage policies to README.md
We basically state that with two registers in an operation r1 and r2 we
prefer to store the output of an instruction in r1. In the case where
the operation is not commutative we prefer ordering the arguments so if
the operation were written down on paper r1 would come before r2 e.g. r1
- r2 is correct, while r2 - r1 is not correct.