Add Normalised() method to Converter interface
This method returns a normalised version of the text input. It also
indicates whether the text was changed when normalised.

This is used by the text-changed listener. The application here is very
declicate because if the text-changed listener finds that the normalised
value is different from the set text it will re-set the text with the
normalised value. This is a recursive and will re-trigger the
text-changed listener, so avoiding infinite recursive loops is very
important.

The UI reason we introduced this is to allow the user to enter a "-"
character at the end of a number input and have this make the number
negative. If the number is already negative this will cause the number
to become positive. We strip out the trailing "-" character so we can
display the value in the conventional way. This makes testing negative
values quite a lot easier.