Add name field to parameter
The parameter struct now includes the name of the parameter. This can be
used for updating the UI and also allowing us to more clearly see the
names of parameters inside demo structs.

Before this the parameters used in demo structs had their names attached
via comments. This is ok - but it would be nicer to attach them directly
to the parameter, and display this in the UI.

This is a somewhat early addition. It is currently unused and only set
by a variety of unused constructors.