Parameter does not expose its Converter
Each parameter instance has a converter field. This allows us to support
converting text from the ui into bytes for our assembly functions and
back out again to the ui.

The Converter interface is _very_ stateless. It only processes the data
you give it and returns an answer, retaining no state. However as the
Parameter struct will store state in the near future we don't want
external code to directly access the Converter, but instead to pass
their data through the Parameter (which delegates the actual conversion
to the Converter). This will allow the Parameter to capture the data as
it is processed.

Currently the Parameter remains entirely stateless and this is simply a
small step towards allowing the Parameter to store state.