Getting/setting parameter data in Parameter itself
We now set the data in the parameter itself. We also take the data from
the parameter when setting UI values.

This commit represents a half-way implementation. The assembly functions
for each demo still take raw bytes and convert them to arrays for
fuction arguments.

The next step is to enable a Parameter to generate typed data slices,
like []uin32 []float64 etc. This will allow us to begin migrating
assembly functions across to use the parameter data instead of using the
raw bytes passed in. Once all demos are converted to using parameters
instead of the raw bytes we can clean up the raw bytes and eliminate
this from the code base.