Setting default parameters for all inputs
A lot of the time it is very tedious to set all the inputs for an
instruction demo. So we will try to set the inputs with unique values
automatically. Obviously things like registers full of bit-masks or
indices will have potentially chaotic results, but I think it will still
be overall much faster to assess and understand an instruction demo.