Parameter.Kind() -> GoType()
This method now returns a more accurate go type declaration for the type
of the parameter.

A few assumptions are made.

If the total bit width is the same as the width of each part then the
type is a straight instance of the type captured by the converter. e.g.

int32 or uint64, float64 etc.

This means that an array of size 1 will get converted into the non-array
type. We don't use size 1 arrays in practice so this is _ok_ enough.

If the total bit width is larger than the width of each part then we
assume it is a pointer to an array (array values are not allowed). e.g.

*[8]uint32 etc.