Add dev_notes/README.md
This is a short description of how my approach to generating code for
instructions changed over time. Also describes my initial use of AI and
how it turned out to be too slow and mentally fatiguing (both surprising
results to me) to use AI for this highly repetitive code-gen.

Building a more capable deterministic code-gen tool was far more
valuable and much faster than building a code skeleton and having AI
fill in the implementation.